+++ Jvedio-WPF/Jvedio/Core/Scraper/Models/ScrapedActor.cs
@@ -1,0 +1,11 @@
+namespace Jvedio.Core.Scraper.Models
+{
+    public class ScrapedActor
+    {
+        public string ActorId { get; set; }
+
+        public string Name { get; set; }
+
+        public string AvatarUrl { get; set; }
+    }
+}